ai assistant/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.9  # PostgreSQL adapter
python-dotenv==1.0.0
openai>=1.12.0
google-generativeai>=0.3.0
requests>=2.31.0
twilio>=8.10.0  # For SMS and WhatsApp
djangorestframework-simplejwt==5.3.0
Pillow>=10.2.0  # Using newer version with pre-built wheels for Windows
setuptools>=65.0.0  # Required for pkg_resources (needed by simplejwt)
openpyxl>=3.1.0  # For Excel file support (.xlsx)
stripe>=7.0.0  # For payment processing

